+++ lib/ts/types.ts
@@ -222,6 +222,16 @@ export type FormFieldBaseConfig = {
      * placeholder of the input field.
      */
     placeholder?: string;
+
+    /*
+     *Ability to provide default value to input field.
+     */
+    getDefaultValue?: () => string;
+
+    /*
+     * Error message for non optional field.
+     */
+    nonOptionalErrorMsg?: string;
 };
 
 export type FormField = FormFieldBaseConfig & {
@@ -275,6 +285,11 @@ export type NormalisedFormField = {
     optional: boolean;
 
     /*
+     * Error message for non optional field.
+     */
+    nonOptionalErrorMsg?: string;
+
+    /*
      * Autocomplete input.
      */
     autoComplete?: string;
@@ -283,6 +298,11 @@ export type NormalisedFormField = {
      * Moves focus to the input element when it mounts
      */
     autofocus?: boolean;
+
+    /*
+     *Ability to provide default value to input field.
+     */
+    getDefaultValue?: () => string;
 };
 
 export type ReactComponentClass<P = any> = ComponentClass<P, any> | ((props: P) => JSX.Element);